Add condition to push NuGet packages only from the main branch (#215)
This pull request introduces a small but important change to the
`.azdo/ci.yaml` pipeline configuration. The change ensures that the
"Push NuGet Packages" step only runs when the build succeeds and the
source branch is `main`.

* Added a condition to the "Push NuGet Packages" pipeline task so that
it only executes on successful builds from the `main` branch
(`.azdo/ci.yaml`).